+++ schema/deploy/view_ciip_production.sql
@@ -19,13 +19,13 @@ BEGIN;
     select
        row_number() over (Partition by true) as row_id,
        x.id as application_id,
-       x.facility_data ->> 'bcghgid' as bcghgid,
-       x.production_data ->> 'quantity' as quantity,
+       (x.facility_data ->> 'bcghgid')::numeric as bcghgid,
+       (x.production_data ->> 'quantity')::numeric as quantity,
        x.production_data ->> 'product' as product,
        x.production_data ->> 'fuel_units' as fuel_units,
        x.production_data ->> 'comments' as comments,
        x.production_data ->> 'associated_emissions' as associated_emissions,
-       x.production_data ->> 'attributable_fuel_percentage' as attributable_fuel_percentage
+       (x.production_data ->> 'attributable_fuel_percentage')::numeric as attributable_fuel_percentage
     from x
  );
